backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.11.0
httpx>=0.28.0
beautifulsoup4>=4.13.0
lxml>=5.4.0
python-dateutil>=2.9.0
